+++ modules/Assets/cli.php
@@ -1,0 +1,7 @@
+<?php
+
+if (!isset($cli, $app) || PHP_SAPI !== 'cli') {
+    return;
+}
+
+$cli->add(new Assets\Command\Thumbhash($app));