+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add src setting for slect and tag fields to load options from internal sources
 - Sanitize space name during installation
 - Multiselect & delete of assets
+- Add thumbhash to image assets on upload
 
 ## 2.6.3 (2023-08-15)